+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ###Version: 0.14.4
 - ```fix``` Squashed some bugs.
+- ```fix``` Import/Export fixes.
 
 ###Version: 0.14.3
 - ```new``` Added Chinese Traditional translations.